requirements.txt
streamlit>=1.35.0
groq>=0.8.0
endee>=0.1.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
